Fix backend router layering
