+++ backend/tests/unit/test_firmware_pagination.py
@@ -24,11 +24,13 @@ sys.modules.setdefault('google.auth', MagicMock())
 sys.modules.setdefault('google.auth.transport.requests', MagicMock())
 
 from routers.firmware import (
-    get_omi_github_releases,
     FIRMWARE_TAG_PATTERN,
-    MAX_PAGES,
-    _parse_firmware_version,
     _find_candidate_releases,
+    _parse_firmware_version,
+)
+from utils.github_releases import (
+    get_omi_github_releases,
+    MAX_PAGES,
 )
 
 
@@ -104,9 +106,9 @@ class TestPagination:
         mock_client.__aenter__ = AsyncMock(return_value=mock_client)
         mock_client.__aexit__ = AsyncMock(return_value=False)
 
-        with patch('routers.firmware.get_generic_cache', return_value=None), patch(
-            'routers.firmware.set_generic_cache'
-        ) as mock_set_cache, patch('routers.firmware.httpx.AsyncClient', return_value=mock_client), patch.dict(
+        with patch('utils.github_releases.get_generic_cache', return_value=None), patch(
+            'utils.github_releases.set_generic_cache'
+        ) as mock_set_cache, patch('utils.github_releases.httpx.AsyncClient', return_value=mock_client), patch.dict(
             'os.environ', {'GITHUB_TOKEN': 'test-token'}
         ):
 
@@ -139,9 +141,9 @@ class TestPagination:
         mock_client.__aenter__ = AsyncMock(return_value=mock_client)
         mock_client.__aexit__ = AsyncMock(return_value=False)
 
-        with patch('routers.firmware.get_generic_cache', return_value=None), patch(
-            'routers.firmware.set_generic_cache'
-        ), patch('routers.firmware.httpx.AsyncClient', return_value=mock_client), patch.dict(
+        with patch('utils.github_releases.get_generic_cache', return_value=None), patch(
+            'utils.github_releases.set_generic_cache'
+        ), patch('utils.github_releases.httpx.AsyncClient', return_value=mock_client), patch.dict(
             'os.environ', {'GITHUB_TOKEN': 'test-token'}
         ):
 
@@ -166,9 +168,9 @@ class TestPagination:
         mock_client.__aenter__ = AsyncMock(return_value=mock_client)
         mock_client.__aexit__ = AsyncMock(return_value=False)
 
-        with patch('routers.firmware.get_generic_cache', return_value=None), patch(
-            'routers.firmware.set_generic_cache'
-        ), patch('routers.firmware.httpx.AsyncClient', return_value=mock_client), patch.dict(
+        with patch('utils.github_releases.get_generic_cache', return_value=None), patch(
+            'utils.github_releases.set_generic_cache'
+        ), patch('utils.github_releases.httpx.AsyncClient', return_value=mock_client), patch.dict(
             'os.environ', {'GITHUB_TOKEN': 'test-token'}
         ):
 
@@ -184,7 +186,7 @@ class TestCacheBehavior:
     @pytest.mark.asyncio
     async def test_cached_empty_list_is_cache_hit(self):
         """An empty list in cache should NOT trigger a re-fetch."""
-        with patch('routers.firmware.get_generic_cache', return_value=[]) as mock_get:
+        with patch('utils.github_releases.get_generic_cache', return_value=[]) as mock_get:
             result = await get_omi_github_releases("test_key", tag_filter=FIRMWARE_TAG_PATTERN)
 
         assert result == []
@@ -202,9 +204,9 @@ class TestCacheBehavior:
         mock_client.__aenter__ = AsyncMock(return_value=mock_client)
         mock_client.__aexit__ = AsyncMock(return_value=False)
 
-        with patch('routers.firmware.get_generic_cache', return_value=None), patch(
-            'routers.firmware.set_generic_cache'
-        ) as mock_set, patch('routers.firmware.httpx.AsyncClient', return_value=mock_client), patch.dict(
+        with patch('utils.github_releases.get_generic_cache', return_value=None), patch(
+            'utils.github_releases.set_generic_cache'
+        ) as mock_set, patch('utils.github_releases.httpx.AsyncClient', return_value=mock_client), patch.dict(
             'os.environ', {'GITHUB_TOKEN': 'test-token'}
         ):
 
@@ -241,9 +243,9 @@ class TestLastKnownGoodFallback:
         mock_client.__aenter__ = AsyncMock(return_value=mock_client)
         mock_client.__aexit__ = AsyncMock(return_value=False)
 
-        with patch('routers.firmware.get_generic_cache', side_effect=fake_get), patch(
-            'routers.firmware.set_generic_cache'
-        ) as mock_set, patch('routers.firmware.httpx.AsyncClient', return_value=mock_client), patch.dict(
+        with patch('utils.github_releases.get_generic_cache', side_effect=fake_get), patch(
+            'utils.github_releases.set_generic_cache'
+        ) as mock_set, patch('utils.github_releases.httpx.AsyncClient', return_value=mock_client), patch.dict(
             'os.environ', {'GITHUB_TOKEN': 'test-token'}
         ):
 
@@ -273,9 +275,9 @@ class TestLastKnownGoodFallback:
         mock_client.__aenter__ = AsyncMock(return_value=mock_client)
         mock_client.__aexit__ = AsyncMock(return_value=False)
 
-        with patch('routers.firmware.get_generic_cache', side_effect=fake_get), patch(
-            'routers.firmware.set_generic_cache'
-        ), patch('routers.firmware.httpx.AsyncClient', return_value=mock_client), patch.dict(
+        with patch('utils.github_releases.get_generic_cache', side_effect=fake_get), patch(
+            'utils.github_releases.set_generic_cache'
+        ), patch('utils.github_releases.httpx.AsyncClient', return_value=mock_client), patch.dict(
             'os.environ', {'GITHUB_TOKEN': 'test-token'}
         ):
 
@@ -296,9 +298,9 @@ class TestLastKnownGoodFallback:
         mock_client.__aenter__ = AsyncMock(return_value=mock_client)
         mock_client.__aexit__ = AsyncMock(return_value=False)
 
-        with patch('routers.firmware.get_generic_cache', side_effect=fake_get), patch(
-            'routers.firmware.set_generic_cache'
-        ), patch('routers.firmware.httpx.AsyncClient', return_value=mock_client), patch.dict(
+        with patch('utils.github_releases.get_generic_cache', side_effect=fake_get), patch(
+            'utils.github_releases.set_generic_cache'
+        ), patch('utils.github_releases.httpx.AsyncClient', return_value=mock_client), patch.dict(
             'os.environ', {'GITHUB_TOKEN': 'test-token'}
         ):
 
@@ -320,9 +322,9 @@ class TestLastKnownGoodFallback:
         mock_client.__aenter__ = AsyncMock(return_value=mock_client)
         mock_client.__aexit__ = AsyncMock(return_value=False)
 
-        with patch('routers.firmware.get_generic_cache', return_value=None), patch(
-            'routers.firmware.set_generic_cache'
-        ) as mock_set, patch('routers.firmware.httpx.AsyncClient', return_value=mock_client), patch.dict(
+        with patch('utils.github_releases.get_generic_cache', return_value=None), patch(
+            'utils.github_releases.set_generic_cache'
+        ) as mock_set, patch('utils.github_releases.httpx.AsyncClient', return_value=mock_client), patch.dict(
             'os.environ', {'GITHUB_TOKEN': 'test-token'}
         ):
